Adding job scripts for perlmutter

+++ barry/fitter.py
@@ -177,7 +177,7 @@ class Fitter(object):
         # to the python script on the HPC
         return self.is_local() or (len(sys.argv) == 2 and sys.argv[1] == "plot")
 
-    def fit(self, file):
+    def fit(self, file, index=None):
         num_concurrent = self.get_num_concurrent()
 
         num_jobs = self.get_num_jobs()
@@ -189,17 +189,22 @@ class Fitter(object):
             self.logger.info("Running locally on the 0th index.")
             self._run_fit(0, 0)
         elif self.is_interactive():
-            if os.path.exists(self.temp_dir):
-                if self.remove_output:
-                    self.logger.info("Deleting %s" % self.temp_dir)
-                    shutil.rmtree(self.temp_dir)
-            hpc = get_hpc()
-            filename = write_jobscript_slurm(
-                file, name=os.path.basename(file), num_tasks=self.get_num_jobs(), num_concurrent=num_concurrent, delete=False, hpc=hpc
-            )
-            self.logger.info("Running batch job at %s" % filename)
-            config = get_config()
-            os.system(f"{config['hpc_submit_command']} {filename}")
+            if index is None:
+                if os.path.exists(self.temp_dir):
+                    if self.remove_output:
+                        self.logger.info("Deleting %s" % self.temp_dir)
+                        shutil.rmtree(self.temp_dir)
+                hpc = get_hpc()
+                filename = write_jobscript_slurm(
+                    file, name=os.path.basename(file), num_tasks=self.get_num_jobs(), num_concurrent=num_concurrent, delete=False, hpc=hpc
+                )
+                self.logger.info("Running batch job at %s" % filename)
+                config = get_config()
+                os.system(f"{config['hpc_submit_command']} {filename}")
+            else:
+                mi, wi = self._get_indexes_from_index(index)
+                self.logger.info("Running model_dataset %d, walker number %d" % (mi, wi))
+                self._run_fit(mi, wi)
         else:
             if len(sys.argv) == 1:
                 # if launching the job for the first time
